Refactors tests to use traits.
