+++ domain_alias/src/Tests/DomainAliasTestBase.php
@@ -2,13 +2,15 @@
 
 namespace Drupal\domain_alias\Tests;
 
-use Drupal\domain\DomainInterface;
 use Drupal\domain\Tests\DomainTestBase;
+use Drupal\Tests\domain_alias\Functional\DomainAliasTestTrait;
 
 /**
  * Base class and helper methods for testing domain aliases.
  */
 abstract class DomainAliasTestBase extends DomainTestBase {
+
+  use DomainAliasTestTrait;
 
   /**
    * Modules to enable.
@@ -22,45 +24,6 @@ abstract class DomainAliasTestBase extends DomainTestBase {
    */
   protected function setUp() {
     parent::setUp();
-  }
-
-  /**
-   * Creates an alias for testing.
-   *
-   * @param \Drupal\domain\DomainInterface $domain
-   *   A domain entity.
-   * @param string $pattern
-   *   An optional alias pattern.
-   * @param int $redirect
-   *   An optional redirect (301 or 302).
-   * @param int $environment
-   *   An optional environment string.
-   * @param boolean $save
-   *   Whether to save the alias or return for validation.
-   *
-   * @return \Drupal\domain_alias\Entity\DomainAlias
-   *   A domain alias entity.
-   */
-  public function domainAliasCreateTestAlias(DomainInterface $domain, $pattern = NULL, $redirect = 0, $environment = 'default', $save = TRUE) {
-    if (empty($pattern)) {
-      $pattern = '*.' . $domain->getHostname();
-    }
-    $values = array(
-      'domain_id' => $domain->id(),
-      'pattern' => $pattern,
-      'redirect' => $redirect,
-      'environment' => $environment,
-    );
-    // Replicate the logic for creating machine_name patterns.
-    // @see ConfigBase::validate()
-    $machine_name = strtolower(preg_replace('/[^a-z0-9_]/', '_', $values['pattern']));
-    $values['id'] = str_replace(array('*', '.', ':'), '_', $machine_name);
-    $alias = \Drupal::entityTypeManager()->getStorage('domain_alias')->create($values);
-    if ($save) {
-      $alias->save();
-    }
-
-    return $alias;
   }
 
 }
